feat(cli): implement CLI modernization foundation
- Move legacy CLI code to src/legacy/ for reference
- Create clean 3-layer CLI architecture (api/human/workflow)
- Implement comprehensive configuration system with profiles
- Add structured error handling with proper conversions
- Create connection management foundation (stub for now)
- Support profile listing and configuration validation
- Maintain backward compatibility with existing config format

This establishes the core foundation for the modernized CLI.
Next PR will implement the raw API layer.
