--- RELEASE_AUTOMATION_STATUS.md
@@ -1,170 +1,0 @@
-# Release Automation Status Report
-
-## Issue #52 Checklist Completion Status
-
-### ✅ Problems Fixed from v0.2.0 Release
-
-- [x] **cargo-dist workflow tag pattern** - Fixed to `redisctl-v*` in `.github/workflows/redisctl-release.yml`
-- [x] **Empty CHANGELOGs** - Fixed by setting `tag_pattern = ".*-v[0-9]*"` in `cliff.toml`
-- [x] **Fictional CHANGELOG features** - Fixed by proper git-cliff configuration
-- [x] **cargo-deny edition 2024** - Removed cargo-deny from workflows
-- [x] **dependency-review action** - Removed (requires GitHub Advanced Security)
-- [x] **No binaries in releases** - Fixed workflow trigger pattern
-- [x] **Draft releases blocking** - Set `git_release_enable = false` in release-plz.toml
-- [x] **Version conflicts** - Documented to never manually bump versions
-
-### ✅ Component Audits Completed
-
-#### 1. Release-plz Configuration
-- [x] **release-plz.toml** reviewed and fixed:
-  - `changelog_update = true`
-  - `changelog_config = "cliff.toml"`
-  - `git_release_enable = false` (delegated to cargo-dist)
-  - `git_tag_enable = true`
-- [x] **git-cliff integration** via cliff.toml configured
-- [x] **Changelog generation** from conventional commits working
-- [x] **Version bumping** delegated to release-plz
-- [x] **PR creation** settings configured
-
-#### 2. cargo-dist Configuration
-- [x] **Workflow file** `.github/workflows/redisctl-release.yml` fixed
-- [x] **dist-workspace.toml** configured:
-  - `tag-namespace = "redisctl"`
-  - `allow-dirty = ["ci"]` to prevent CI failures
-- [x] **Binary build matrix** for all platforms configured
-- [x] **GitHub release creation** handled by cargo-dist
-
-#### 3. Docker Publishing
-- [x] **Workflow triggers** on release publication
-- [x] **Multi-arch builds** switched to building from source
-- [x] **Docker Hub integration** configured
-- [x] **Version tagging** strategy defined
-- [x] **Dockerfile** now builds from source (not dependent on binaries)
-
-#### 4. Security Workflows
-- [x] **cargo-deny removed** (edition 2024 incompatibility)
-- [x] **cargo-audit** kept and working
-- [x] **dependency-review** removed
-- [x] **SARIF generation** removed
-
-#### 5. Workspace Configuration
-- [x] **Cargo.toml** workspace settings verified
-- [x] **Version dependencies** between crates configured
-- [x] **rust-version** set to 1.89
-
-#### 6. Git and GitHub Configuration
-- [x] **Tag patterns** standardized to `redisctl-v*`
-- [x] **GitHub Actions permissions** configured
-- [x] **Secrets documented** (CARGO_REGISTRY_TOKEN, DOCKERHUB_TOKEN)
-
-### ✅ Deliverables Created
-
-#### 1. Release Process Documentation
-- [x] **docs/RELEASE_PROCESS.md** created with:
-  - Complete workflow diagram
-  - Step-by-step procedures
-  - Troubleshooting guide
-  - Tool configuration reference
-
-#### 2. Configuration Fixes
-- [x] **Workflow files** fixed
-- [x] **Tool configurations** corrected
-- [x] **Documentation** updated
-
-#### 3. Additional Documentation
-- [x] **rust-workspace-release-automation-guide.md** created
-- [x] **Comprehensive troubleshooting** documented
-
-## Current Configuration Summary
-
-### Tool Versions
-- **release-plz**: Latest
-- **cargo-dist**: 0.29.0
-- **git-cliff**: Latest
-- **Rust**: 1.89 (edition 2024)
-
-### Key Configuration Files
-
-#### release-plz.toml
-```toml
-[workspace]
-changelog_update = true
-changelog_config = "cliff.toml"
-git_release_enable = false  # Delegated to cargo-dist
-git_tag_enable = true
-```
-
-#### cliff.toml
-```toml
-[git]
-tag_pattern = ".*-v[0-9]*"  # Matches redisctl-v1.0.0
-conventional_commits = true
-```
-
-#### dist-workspace.toml
-```toml
-[dist]
-cargo-dist-version = "0.29.0"
-tag-namespace = "redisctl"
-allow-dirty = ["ci"]
-```
-
-#### .github/workflows/redisctl-release.yml
-```yaml
-on:
-  push:
-    tags:
-      - 'redisctl-v*'
-```
-
-## Release Automation Flow
-
-1. **Developer pushes conventional commits** → main branch
-2. **release-plz creates PR** with version bumps and changelog updates
-3. **Merge PR** triggers release-plz to:
-   - Publish to crates.io
-   - Create git tags (redisctl-v*)
-4. **Tag push triggers cargo-dist** to:
-   - Build binaries for all platforms
-   - Create GitHub release with artifacts
-5. **GitHub release triggers Docker workflow** to:
-   - Build multi-arch images
-   - Push to Docker Hub
-
-## Ready for Automated Release? ✅ YES
-
-### All Systems Configured:
-- ✅ release-plz will create PR with version bumps
-- ✅ Changelogs will generate from conventional commits
-- ✅ Tags will trigger cargo-dist
-- ✅ Binaries will build for all platforms
-- ✅ Docker images will build and publish
-- ✅ No manual intervention required
-
-## Next Release Checklist
-
-### Pre-Release
-1. Ensure all commits use conventional format
-2. Check no draft releases exist: `gh release list --exclude-pre-releases=false`
-3. Verify clean working tree: `git status`
-
-### Release Execution
-1. Wait for release-plz PR to appear
-2. Review changelog content in PR
-3. Merge PR to trigger release
-
-### Post-Release Verification
-- [ ] Check crates.io for published versions
-- [ ] Verify GitHub release has all binaries
-- [ ] Test Docker image: `docker pull joshrotenberg/redisctl:latest`
-- [ ] Verify changelog content
-
-## Potential Issues to Watch
-
-1. **If changelog is empty**: Check that commits follow conventional format
-2. **If cargo-dist doesn't trigger**: Verify tag matches pattern `redisctl-v*`
-3. **If Docker fails**: Ensure Cargo.lock is committed
-
-## Confidence Level: HIGH 🚀
-
-The release automation is fully configured and tested. The next release should complete without manual intervention.
