+++ crates/redisctl/legacy/api_tests.rs
