+++ crates/redisctl/src/connection.rs
@@ -1,0 +1,15 @@
+//! Connection management for Redis Cloud and Enterprise clients (stub for now)
+
+use crate::config::Config;
+
+/// Connection manager for creating authenticated clients (stub implementation)
+pub struct ConnectionManager {
+    pub config: Config,
+}
+
+impl ConnectionManager {
+    /// Create a new connection manager with the given configuration
+    pub fn new(config: Config) -> Self {
+        Self { config }
+    }
+}
