+++ crates/redisctl/src/legacy/cloud_bin.rs
