+++ crates/redisctl/src/legacy/commands/api.rs
