+++ crates/redisctl/src/legacy/commands/auth.rs
