+++ crates/redisctl/src/legacy/commands/enterprise.rs
