+++ crates/redisctl/src/legacy/commands/mod.rs
