+++ crates/redisctl/src/legacy/commands/profile.rs
