+++ crates/redisctl/src/legacy/enterprise_bin.rs
