+++ crates/redisctl/src/legacy/router.rs
